packages/compost/pyproject.toml
[tool.poetry]
name = "fuit-es"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pygame = "^2.6.1"
noise = "^1.2.2"
numpy = "^2.1.2"
matplotlib = "^3.5.3"
scikit-image = "^0.24.0"
scikit-learn = "^1.5.2"
flask = "^3.1.0"
opencv-python = "^4.10.0.84"
shapely = "^2.0.6"
pymunk = "^6.9.0"
pyaml = "^25.1.0"
librosa = "^0.11.0"
ipywidgets = "^8.1.7"
plotly = "^6.1.1"
nbformat = "^5.10.4"
asteroid = "^0.7.0"
demucs = "^4.0.1"
soundfile = "^0.13.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
